backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pinecone-client==5.0.1
boto3==1.35.36
pydantic==2.10.3

